### 2022-02-17 #4179 @whedon
[PRE REVIEW]: lintegrate: A C/Python numerical integration library for working in log-space
**Submitting author:** @mattpitkin (<a href="http://orcid.org/0000-0003-4548-526X">Matthew Pitkin</a>)
**Repository:** <a href="https://github.com/mattpitkin/lintegrate" target ="_blank">https://github.com/mattpitkin/lintegrate</a>
**Version:** v0.1.11
