### 2022-02-17 #4179 @danielskatz
👋 @mattpitkin - thanks for your submission.  Due to the relatively small amount of original code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS.  You should hear back in a week or so.
