### 2022-02-27 #4179 @arfon
@mattpitkin – could you explain a little more about the lineage/heritage of this code? I see a fair chunk of C code in the `src` folder which seems to be from a number of authors, and then the Python bindings/wrapper code?

Wrappers of other code generally aren't in scope for JOSS (especially if the code being wrapped isn't written by the author) but I'm not sure if I'm reading this correctly here.
