### 2022-02-28 #4179 @mattpitkin
@arfon Sure. For a bit of background, a while back (for the work I was doing for [this paper](https://ui.adsabs.harvard.edu/abs/2018PhRvD..98f3001P/abstract)) I was wanting to numerically integrate a likelihood for which it only made sense numerically to work in log-space. I couldn't find any existing routines that would do this, but I did still want to make as much use as I could of existing numerical integration routines rather than coding everything from scratch. I decided that I'd develop these routines as a C-library and then create a cython wrapper to them. For the C-library, I've used several existing numerical integration routines from GSL (mainly authored by Brian Gough and Pedro Gonnet) and edited them so that all the summations work in log-space. While this mainly consisted of replacing sums with the [`logaddexp`](https://github.com/mattpitkin/lintegrate/blob/master/src/logadd.c) and [`logsubexp`](https://github.com/mattpitkin/lintegrate/blob/master/src/logsub.c) functions that I wrote, it was a not entirely trivial process as various error checking an tolerance calculations also had to be changed (so a considerable fraction of the lines in the [`lintegrate_cquad.c`](https://github.com/mattpitkin/lintegrate/blob/master/src/lintegrate_cquad.c), [`lintegrate_qag.c`](https://github.com/mattpitkin/lintegrate/blob/master/src/lintegrate_qag.c) and [`lintegrate_qng.c`](https://github.com/mattpitkin/lintegrate/blob/master/src/lintegrate_qng.c), particularly in the main functions rather than various helper functions, are my changes). Additionally, my C-library includes the new "split" [functions](https://github.com/mattpitkin/lintegrate/blob/master/src/lintegrate_split.c) that allow the integral to be broken up into sub-intervals in the abscissa variable.

The Cython wrapper was all my work. This also required significant changes to the C-code, so that when used from Python the library could take in Python functions. In addition to wrapper to the main numerical routines (done in a way to be consistent with the SciPy integration routines) it also provides the `logtrapz` function for just performing trapezoid-rule integration in log-space.

In additional to my own use of the code, it has also been used by others in publications for machine learning applications.
