### 2022-03-04 #4179 @jbytecode
👋👋👋 Dear @Het-Shah  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**lintegrate: A C/Python numerical integration library for working in log-space**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4179).
