### 2022-03-04 #4179 @Het-Shah
> 👋👋👋 Dear @Het-Shah 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**lintegrate: A C/Python numerical integration library for working in log-space**". You can find more information at the top of this Github issue (#4179).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.

I am fine reviewing. Thank you!
