### 2022-03-07 #4231 @editorialbot
[REVIEW]: lintegrate: A C/Python numerical integration library for working in log-space
**Submitting author:** @mattpitkin (<a href="http://orcid.org/0000-0003-4548-526X">Matthew Pitkin</a>)
**Repository:** https://github.com/mattpitkin/lintegrate
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.11
