### 2022-03-24 mattpitkin/lintegrate#22 @Het-Shah
Review for JOSS submission #4231
Overall the submission [(#4231)](https://github.com/openjournals/joss-reviews/issues/4231) seems to be good to go. Some comments -

- Contributor guidelines are missing. Kindly add those to the repository.
- Paper seems to have some grammatical errors. I would recommend running a grammar check on the paper. I am mentioning some of them below:
  - Line 6 situation -> situations
  - Line 29 "Instead you must work with the natural logarithm of the likelihood" -> "Instead, it would help to work with the natural logarithm of the likelihood."
- Are there any other libraries/repositories which attempt to solve a similar problem? If not, can this be mentioned in the paper?
