### 2022-03-24 00791be
Some minor grammatical/typo fixes as suggested in #22
