### 2022-03-24 #4231 @mattpitkin
@Het-Shah Thanks for your review. I'll add my responses here:

> * Contributor guidelines are missing. Kindly add those to the repository.

I have contribution guidelines on the docs page [here](https://lintegrate.readthedocs.io/en/latest/installation.html#contributing) and I've now added a `CONTRIBUTING.md` file to the repo with the same information https://github.com/mattpitkin/lintegrate/commit/c368238e2a89c7eeeea34ccd0ea6a39deeba3751.

> * Paper seems to have some grammatical errors. I would recommend running a grammar check on the paper. I am mentioning some of them below:
>   * Line 6 situation -> situations
>   * Line 29 "Instead you must work with the natural logarithm of the likelihood" -> "Instead, it would help to work with the natural logarithm of the likelihood."

Thanks for spotting these. I've made some changes in https://github.com/mattpitkin/lintegrate/commit/00791be10c8014d2f84f8da3a84b41a2e1ead209.

> * Are there any other libraries/repositories which attempt to solve a similar problem? If not, can this be mentioned in the paper?

There are no other libraries that address this as far as I am aware.
