### 2022-03-24 88fee9d
Attempt to fix mac osx builds by requiring clang>=12.0.1
