### 2022-03-24 cec112b
add c-compiler rather than compilers to mac osx build as compilers also does various other unnecessary comilers
