### 2022-04-06 96d7afb
Minor fix to pxd file to move header file imports back into pyx file. Also get rid of Numpy version requirement
