+++ lintegrate/__init__.py
@@ -2,7 +2,7 @@ import os
 
 from .lintegrate import *
 
-__version__ = "0.1.12"
+__version__ = "0.1.13"
 
 
 def get_include():
