+++ setup.cfg
@@ -29,7 +29,7 @@ setup_requires =
     numpy
     setuptools_scm
 install_requires = 
-    numpy>=1.21
+    numpy
 packages = find:
 
 [options.package_data]
