### 2022-04-18 #4231 @jbytecode
👋👋👋 Dear @jakryd  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**lintegrate: A C/Python numerical integration library for working in log-space**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4231).

The 2nd reviewer of this submission does not respond to our requests, so I am sending this email for completing the halfly finished reviewing job. Whenever you accept the invitation, we will generate a new set of reviewing tasks.
