### 2022-04-19 #4231 @jbytecode
@jakryd - Thank you so much for accepting our invitation. Could you please type

`@editorialbot generate my checklist`

to generate your checklist? In this list, there will be 20 reviewing item. Whenever you complete them, you can check the corresponding item. You can also interact with the authors, other reviewers, and the editor anytime you need.

Do not hesitate to ask me anything.

Thank you in advance.
