### 2022-04-20 a41867d
Update paper.md
add availability and development section
