### 2022-04-20 #4231 @mattpitkin
For information, I have added an "Availability and development" section to the paper containing:

> lintegrate is open source code, released under a GPL-3.0 license. The source code is publically hosted on [GitHub](https://github.com/mattpitkin/lintegrate), while release versions of the Python library are availble through [PyPI](https://pypi.org/project/lintegrate/) and [conda-forge](https://anaconda.org/conda-forge/lintegrate). Documentation of the C library functions and Python interface can be found on [Read the Docs](https://lintegrate.readthedocs.io/en/latest/), including the API and example usage.
>
> Contributions to the code are welcome via the GitHub [issues tracker](https://github.com/mattpitkin/lintegrate/issues) or [Pull Requests](https://github.com/mattpitkin/lintegrate/pulls). Continuous integration is set up to run a test suite on the code when any changes are made.
