### 2022-04-20 62a651d
Update paper.md
fix typo
