### 2022-04-21 #4231 @jbytecode
> @jakryd - Thank you so much for accepting our invitation. Could you please type
>
> `@editorialbot generate my checklist`
>
> to generate your checklist? In this list, there will be 20 reviewing items. Whenever you complete them, you can check the corresponding item. You can also interact with the authors, other reviewers, and the editor anytime you need.
>
> Do not hesitate to ask me anything.
>
> Thank you in advance.

@jakryd - do you receive the notifications? Could you please start your review by generating the checklist as mentioned above?
