### 2022-05-18 #4231 @jakryd
I have completed my checkbox. The article and code look very well. I have two minor suggestions for @mattpitkin to consider. (1) I would capitalize "Proceedings of the 36th international conference on machine learning" in the references to be consistent. (2) "For C codes, there are a range of numerical integration (sometimes called numerical quadrature) functions" -- I would say "there is a range."
