+++ paper/paper.bib
@@ -27,12 +27,12 @@ archivePrefix = {arXiv},
 @InProceedings{Nash:2019,
   title = 	 {Autoregressive Energy Machines},
   author =       {Nash, Charlie and Durkan, Conor},
-  booktitle = 	 {Proceedings of the 36th International Conference on Machine Learning},
+  booktitle = 	 "{Proceedings of the 36th International Conference on Machine Learning}",
   pages = 	 {1735--1744},
   year = 	 {2019},
   editor = 	 {Chaudhuri, Kamalika and Salakhutdinov, Ruslan},
   volume = 	 {97},
-  series = 	 {Proceedings of Machine Learning Research},
+  series = 	 "{Proceedings of Machine Learning Research}",
   month = 	 {09--15 Jun},
   publisher =    {PMLR},
   pdf = 	 {http://proceedings.mlr.press/v97/durkan19a/durkan19a.pdf},
