### 2022-05-19 eda3e18
Corrections for JOSS
