+++ paper/paper.bib
@@ -50,3 +50,12 @@ archivePrefix = {arXiv},
   eprint    = {2102.04426},
   biburl    = {https://dblp.org/rec/journals/corr/abs-2102-04426.bib},
 }
+
+@article{vanrossum1995python,
+  title={{P}ython reference manual},
+  author={vanRossum, Guido},
+  journal={Department of Computer Science [CS]},
+  number={R 9525},
+  year={1995},
+  publisher={CWI}
+}
