### 2022-05-19 #4231 @jbytecode
Dear @mattpitkin

- [ ] I have just sent a pull request that includes some corrections for the paper. Please apply them if you are agreed with it.
- [ ] Please read the paper pdf carefully, one more time.
- [ ] Create a tagged release in the software repository. Report the version here. The version should be in form of `vx.y.z`, e.g. `v1.2.3`.
- [ ] Create a Zenodo archive in https://zenodo.org/
- [ ] The Zenodo archive (with the same name/title with the JOSS paper) should have the correct metadata, e.g. author names, ORCIDs, the title of the archive. Despite it is not a rigid requirement, please set the archive title as the paper title.
- [ ] Please provide the tagged release version id, zenodo archive DOI, zenodo archive url here.

Thank you in advance.
