### 2022-05-19 ba69e47
Update paper.md
Minor changes
