### 2022-05-19 #4231 @mattpitkin
The latest tagged release is [v0.1.13](https://github.com/mattpitkin/lintegrate/releases/tag/v0.1.13), The Zenodo DOI for this release is [10.5281/zenodo.6419039](https://doi.org/10.5281/zenodo.6419039) and the Zenodo archive URL is https://zenodo.org/record/6419039#.YoYIQ3XMKV5.
