### 2022-05-19 4357d6c
Update paper.md
Fix typo
