### 2022-05-19 48b15cb
Update paper.md
another typo fix
