### 2022-05-19 #4231 @mattpitkin
@Het-Shah @jakryd thanks for reviewing the paper/software and @jbytecode that for the editorial input. I've checked over the paper and fixed another couple of minor typos, but am now happy for it to be released.
